Bump bo4e-schema-tool from 0.0.8 to 0.0.9 (use Decimal instead of float) (#223)
* 🚧

* new generate

* WIP

* Add required fields

* Bump schema-Tool to 0.0.9

+++ src/ibims/bo4e/com/auf_abschlagstaffel_pro_ort.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import TYPE_CHECKING, Optional
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -31,15 +32,15 @@ class AufAbschlagstaffelProOrt(BaseModel):
     """
     Version der BO-Struktur aka "fachliche Versionierung"
     """
-    staffelgrenze_bis: Optional[float] = Field(default=None, alias="staffelgrenzeBis", title="Staffelgrenzebis")
+    staffelgrenze_bis: Optional[Decimal] = Field(default=None, alias="staffelgrenzeBis", title="Staffelgrenzebis")
     """
     Oberer Wert, bis zu dem die Staffel gilt.
     """
-    staffelgrenze_von: Optional[float] = Field(default=None, alias="staffelgrenzeVon", title="Staffelgrenzevon")
+    staffelgrenze_von: Optional[Decimal] = Field(default=None, alias="staffelgrenzeVon", title="Staffelgrenzevon")
     """
     Unterer Wert, ab dem die Staffel gilt.
     """
-    wert: Optional[float] = Field(default=None, title="Wert")
+    wert: Optional[Decimal] = Field(default=None, title="Wert")
     """
     Der Wert für den Auf- oder Abschlag.
     """
